app/requirements.txt
requests
tinydb
python-telegram-bot>=12.0.0b1,<13.0.0
